chore: split cli tests in 3 (#16017)
