+++ examples/full-jaffle-shop-demo/dbt/models/customers.yml
@@ -5,7 +5,7 @@ models:
       meta:
         primary_key: customer_id
         spotlight:
-          categories: ["experimental"]
+          categories: [ "experimental" ]
         joins:
           - join: membership
             hidden: true
@@ -14,7 +14,7 @@ models:
           - join: plan
             sql_on: ${membership.plan_id} = ${plan.id}
             relationship: many-to-one
-        ai_hint: 
+        ai_hint:
           - This is a customers table.
           - Contains customer information and demographics.
     description: |
@@ -36,8 +36,11 @@ models:
                 label: Unique customer count
                 description: Total number of customers
                 spotlight:
-                  categories: ["core"]
+                  categories: [ "core" ]
                 ai_hint: Unique customer count.
+        meta:
+          dimension:
+            type: number
       - name: first_name
         description: |
           # First name
@@ -51,25 +54,26 @@ models:
               type: string
               urls:
                 - label: "URL from value"
-                  url:
-                    "https://example.com/company/${value.formatted | url_encode
-                    }"
+                  url: "https://example.com/company/${value.formatted | url_encode }"
                 - label: "URL from row value"
-                  url:
-                    "https://example.com/company/${row.customers.customer_id.raw
-                    | url_encode }"
-                - label: "Invalid URL with bad reference"
-                  url:
-                    "https://example.com/company/${row.customer_id.raw | url_encode
+                  url: "https://example.com/company/${row.customers.customer_id.raw | url_encode
                     }"
+                - label: "Invalid URL with bad reference"
+                  url: "https://example.com/company/${row.customer_id.raw | url_encode }"
+        meta:
+          dimension:
+            type: string
       - name: last_name
         description: Customer's last name. PII.
         config:
           meta:
             dimension:
-              ai_hint: 
+              ai_hint:
                 - Last name of the customer
                 - Contains PII data
+        meta:
+          dimension:
+            type: string
       - name: age
         description: Customer's age
         config:
@@ -82,6 +86,9 @@ models:
               average_age:
                 type: average
                 description: Average age of customers
+        meta:
+          dimension:
+            type: number
       - name: created
         description: Timestamp (UTC) when customer was created
         config:
@@ -93,24 +100,36 @@ models:
                 spotlight:
                   visibility: hide
                 type: max
+        meta:
+          dimension:
+            type: timestamp
       - name: first_order
         description: Date of the customers first order
         config:
           meta:
             dimension:
               hidden: true
+        meta:
+          dimension:
+            type: date
       - name: most_recent_order
         description: Date of the customers most recent order
         config:
           meta:
             dimension:
               hidden: true
+        meta:
+          dimension:
+            type: date
       - name: days_since_last_order
         description: Number of days since the customers last order
         config:
           meta:
             dimension:
               hidden: true
+        meta:
+          dimension:
+            type: number
       - name: days_between_created_and_first_order
         description: >-
           Number of days between the customer being created and making their
@@ -119,15 +138,24 @@ models:
           meta:
             dimension:
               hidden: true
+        meta:
+          dimension:
+            type: number
       - name: number_of_orders
         description: ""
         config:
           meta:
             dimension:
               hidden: true
+        meta:
+          dimension:
+            type: number
       - name: customer_lifetime_value
         description: ""
         config:
           meta:
             dimension:
               hidden: true
+        meta:
+          dimension:
+            type: number
